+++ controller/Changelog.md
@@ -1,3 +1,13 @@
+# [2026.3.0-VALIDATION.1] - 2026-03-12
+
+# Added
+
+- kiosk: Clear cached files if kiosk fails to start repeatedly
+
+# Changed
+
+- driver: Upgrade driver to 2.8.1 to remove irrelevant error log
+
 # [2026.3.0-VALIDATION] - 2026-02-25
 
 ## Added